Refactor: change GetUser -> GetUserByID and create error object where it happen
