+++ internal/core/ports/user_port.go
@@ -9,7 +9,7 @@ import (
 
 type UserRepository interface {
 	Create(user *domain.User) error
-	GetUser(userID uuid.UUID) (*domain.User, error)
+	GetUserByID(userID uuid.UUID) (*domain.User, error)
 	UpdateInformation(userID uuid.UUID, data dto.UserInformationRequestBody) error
 	UpdateUser(user domain.User) error
 	GetUserByEmail(email string) (*domain.User, error)
